Document id int to bigint migration (#3463)
* Custom bigint document_id migration

* Two transactions to migrate document id

+++ server/core/models/content_ad/model.py
@@ -132,7 +132,7 @@ class ContentAd(
     video_asset = models.ForeignKey("VideoAsset", blank=True, null=True, on_delete=models.PROTECT)
     ad_tag = models.TextField(null=True, blank=True)
 
-    document_id = models.IntegerField(null=True, blank=True)
+    document_id = models.BigIntegerField(null=True, blank=True)
     document_features = JSONField(null=True, blank=True)
 
     redirect_id = models.CharField(max_length=128, null=True)
